skypebot_support/requirements.txt
skpy
python-dotenv
icmplib
speedtest-cli
requests
pymssql